// src/app/models/userRegisterResponse.interface.ts
export interface UserRegisterResponseI {
    userId: string;
    createdAt: string;
    emailStatus: string;
    message: string;
    emailExpiresAt: string;
    firstName: string;
    lastName: string;
    email: string;
    country: string;
    address: string;
    dateOfBirth: string;
    zipcode: string;
    password: string;
}